requirements.txt
# requirements.txt

# Para a API
fastapi
uvicorn[standard]

# Para Embeddings e pré-processamento
sentence-transformers
nltk
langchain-text-splitters  # Usaremos para um chunking mais inteligente

# Para o Banco de Dados Postgres + PGVector
psycopg2-binary
pgvector

# Para o Agente (Bônus) - conectar ao LM Studio
openai

# Para variáveis de ambiente (boa prática)
python-dotenv

# Para MCP Server
mcp>=1.0.0
httpx>=0.24.0